+++ xenon-common/src/test/java/com/vmware/xenon/services/common/ReplicationTestService.java
@@ -84,7 +84,7 @@ public class ReplicationTestService extends StatefulService {
         t.documentSelfLink = UUID.randomUUID().toString();
         initState.queryTaskLink = UriUtils.buildUriPath(ServiceUriPaths.CORE_QUERY_TASKS,
                 t.documentSelfLink);
-        sendRequest(Operation.createPost(this, LuceneQueryTaskFactoryService.SELF_LINK).setBody(t));
+        sendRequest(Operation.createPost(this, QueryTaskFactoryService.SELF_LINK).setBody(t));
         startPost.complete();
 
         initState.stringField = UUID.randomUUID().toString();